+++ react/features/base/util/index.js
@@ -1,4 +1,3 @@
-export * from './interceptComponent';
 export * from './loadScript';
 export * from './randomUtil';
 export * from './roomnameGenerator';
